Home

NiLang.jl

NiLang is a reversible eDSL that can run backwards. The motation is to support source to source AD.

Our paper is comming soon.

Welcome for discussion in Julia slack, #autodiff channel.

NiLang.CONJ

NiLang.DIVINT

NiLang.IROT

NiLang.MULINT

NiLang.NEG

NiLang.ROT

NiLang.SWAP

NiLang.XOR

NiLang.arshift

NiLang.plshift

NiLang.prshift

NiLang.rot

NiLang.CONJ — Method.

CONJ(a!) -> a!'

source

NiLang.DIVINT — Method.

DIVINT(a!, b::Integer) -> a!/b, b

source

NiLang.IROT — Method.

IROT(a!, b!, θ) -> ROT(a!, b!, -θ)

source

NiLang.MULINT — Method.

MULINT(a!, b::Integer) -> a!*b, b

source

NiLang.NEG — Method.

NEG(a!) -> -a!

source

NiLang.ROT — Method.

ROT(a!, b!, θ) -> a!', b!', θ

\[\begin{align} {\rm ROT}(a!, b!, \theta) = \begin{bmatrix} \cos(\theta) & - \sin(\theta)\\ \sin(\theta) & \cos(\theta) \end{bmatrix} \begin{bmatrix} a!\\ b! \end{bmatrix}, \end{align}\]

source

NiLang.SWAP — Method.

SWAP(a!, b!) -> b!, a!

source

NiLang.XOR — Method.

XOR(a!, b) -> a! ⊻ b, b

source

NiLang.arshift — Method.

arshift(x, n)

right shift, sign extending.

source

NiLang.plshift — Method.

plshift(x, n)

periodic left shift.

source

NiLang.prshift — Method.

plshift(x, n)

periodic right shift.

source

NiLang.rot — Method.

rot(a, b, θ)

rotate variables a and b by an angle θ

source