API Manual

API Manual

Compiling Tools (Reexported from NiLangCore)

NiLangCore.@anc — Macro.

@anc x = expr

Create an ancilla x with initial value expr,

NiLangCore.@assign — Macro.

@assign a b [invcheck]

Perform the assign a = b in a reversible program. Turn off invertibility check if the invcheck is false.

NiLangCore.@assignback — Macro.

@assignback f(args...) [invcheck]

Assign input variables with output values: args... = f(args...), turn off invertibility error check if the second argument is false.

NiLangCore.@code_interpret — Macro.

@code_interpret ex

Get the interpreted expression of ex.

julia> using MacroTools

julia> prettify(@code_interpret x += exp(3.0))
:(@assignback (PlusEq(exp))(x, 3.0))

NiLangCore.@code_preprocess — Macro.

@code_preprocess ex

Preprocess ex and return the symmetric reversible IR.

julia> using MacroTools

julia> prettify(@code_preprocess if (x < 3, ~) x += exp(3.0) end)
:(if (x < 3, x < 3)
      x += exp(3.0)
  else
  end)

NiLangCore.@code_reverse — Macro.

@code_reverse ex

Get the reversed expression of ex.

julia> @code_reverse x += exp(3.0)
:(x -= exp(3.0))

NiLangCore.@deanc — Macro.

@deanc x = expr

Deallocate ancilla x if x == expr, else throw an InvertibilityError.

NiLangCore.@dual — Macro.

@dual f invf

Define f and invf as a pair of dual instructions, i.e. reverse to each other.

NiLangCore.@fieldview — Macro.

@fieldview fname(x::TYPE) = x.fieldname

Create a function fieldview that can be accessed by a reversible program

julia> struct GVar{T, GT}
           x::T
           g::GT
       end

julia> @fieldview xx(x::GVar) = x.x

julia> chfield(GVar(1.0, 0.0), xx, 2.0)
GVar{Float64,Float64}(2.0, 0.0)

NiLangCore.@i — Macro.

@i function fname(args..., kwargs...) ... end
@i struct sname ... end

Define a reversible function/type.

julia> @i function test(out!, x)
           out! += identity(x)
       end

julia> test(0.2, 0.8)
(1.0, 0.8)

julia> @i struct CVar{T}
           g::T
           x::T
           function CVar{T}(x::T, g::T) where T
               new{T}(x, g)
           end
           function CVar(x::T, g::T) where T
               new{T}(x, g)
           end
           # warning: infered type `T` should be used in `← new` statement only.
           @i function CVar(xx::T) where T
               gg ← zero(xx)
               gg += identity(1)
               xx ← new{T}(gg, xx)
           end
       end

julia> CVar(0.2)
CVar{Float64}(1.0, 0.2)

julia> (~CVar)(CVar(0.2))
0.2

See test/compiler.jl and test/invtype.jl for more examples.

NiLangCore.@icast — Macro.

@icast TDEF1 => TDEF2 begin ... end

Define type cast between TDEF1 and TDEF2. Type TDEF1 or TDEF2 can be T(args...) or x::T like.

julia> struct PVar{T}
           g::T
           x::T
       end

julia> struct SVar{T}
           x::T
           g::T
       end

julia> @icast PVar(g, x) => SVar(x, k) begin
          g → zero(x)
          k ← zero(x)
          k += identity(x)
       end

julia> x = PVar(0.0, 0.5)
PVar{Float64}(0.0, 0.5)

julia> @instr (PVar=>SVar)(x)
SVar{Float64}(0.5, 0.5)

julia> @instr (SVar=>PVar)(x)
PVar{Float64}(0.0, 0.5)

julia> @icast x::Base.Float64 => SVar(x, gg) begin
           gg ← zero(x)
           gg += identity(x)
       end

julia> x = 0.5
0.5

julia> @instr (Float64=>SVar)(x)
SVar{Float64}(0.5, 0.5)

julia> @instr (SVar=>Float64)(x)
0.5

NiLangCore.@instr — Macro.

@instr ex

Execute a reversible instruction.

NiLangCore.@invcheck — Macro.

@invcheck ex
@invcheck x val

Pass the check it if ex is true or x ≈ val.

NiLangCore.@pure_wrapper — Macro.

@pure_wrapper TYPE

Create a reversible wrapper type TYPE{T} <: IWrapper{T} that plays a role of simple wrapper.

```jldoctest; setup=:(using NiLangCore) julia> @pure_wrapper A

julia> A(0.5) A(0.5)

julia> (~A)(A(0.5)) 0.5

julia> -A(0.5) A(-0.5)

julia> A(0.5) < A(0.6) true

NiLangCore.@selfdual — Macro.

@selfdual f

Define f as a self-dual instructions.

NiLangCore.almost_same — Method.

almost_same(a, b; atol=GLOBAL_ATOL[], kwargs...) -> Bool

Return true if a and b are almost same w.r.t. atol.

NiLangCore.check_inv — Method.

check_inv(f, args; kwargs=(), atol::Real=1e-8, verbose::Bool=false)

Return true if f(args..., kwargs...) is reversible.

NiLangCore.chfield — Function.

chfield(x, field, val)

Change a field of an object x.

The field can be a Val type

julia> chfield(1+2im, Val(:im), 5)
1 + 5im

or a function

julia> using NiLangCore

julia> struct GVar{T, GT}
           x::T
           g::GT
       end

julia> @fieldview xx(x::GVar) = x.x

julia> chfield(GVar(1.0, 0.0), xx, 2.0)
GVar{Float64,Float64}(2.0, 0.0)

NiLangCore.isprimitive — Method.

isprimitive(f)

Return true if f is an instruction that can not be decomposed anymore.

NiLangCore.isreflexive — Method.

isreflexive(f)

Return true if a function is self-inverse.

NiLangCore.isreversible — Method.

isreversible(f)

Return true if a function is reversible.

NiLangCore.nargs — Function.

nargs(instr)

Number of arguments as the input of instr.

NiLangCore.nouts — Function.

nouts(instr)

Number of output arguments in the input list of instr.

NiLangCore.tget — Method.

tget(x::Tuple, i)

Get the i-th entry of a tuple.

NiLangCore.unwrap — Method.

unwrap(x)

Unwrap a wrapper instance (recursively) to get the original value.

NiLangCore.value — Method.

value(x)

Get the value from a wrapper instance.

NiLangCore.IWrapper — Type.

IWrapper{T} <: RevType

IWrapper{T} is a wrapper of for data of type T. It will forward >, <, >=, <=, ≈ operations.

NiLangCore.Inv — Type.

Inv{FT} <: Function
Inv(f)

The inverse of a function.

NiLangCore.InvertibilityError — Type.

InvertibilityError <: Exception
InvertibilityError(ex)

The error thrown when a irreversible statement appears in a reversible context.

NiLangCore.MinusEq — Type.

MinusEq{FT} <: Function
MinusEq(f)
⊖(f)

Called when executing out -= f(args...) instruction. See PlusEq for detail.

NiLangCore.Partial — Type.

Partial{FIELD, T, T2} <: IWrapper{T2}

Take a field FIELD without dropping information. This operation can be undone by calling ~Partial{FIELD}.

NiLangCore.PlusEq — Type.

PlusEq{FT} <: Function
PlusEq(f)
⊕(f)

Called when executing out += f(args...) instruction. The following two statements are same

julia> x, y, z = 0.0, 2.0, 3.0
(0.0, 2.0, 3.0)

julia> x, y, z = PlusEq(*)(x, y, z)
(6.0, 2.0, 3.0)

julia> x, y, z = 0.0, 2.0, 3.0
(0.0, 2.0, 3.0)

julia> @instr x += y*z
3.0

julia> x, y, z
(6.0, 2.0, 3.0)

NiLangCore.RevType — Type.

RevType

The base type for reversible types.

NiLangCore.XorEq — Type.

XorEq{FT} <: Function
XorEq(f)
⊙(f)

Called when executing out ⊻= f(args...) instruction. See PlusEq for detail.

NiLangCore.assign_vars — Method.

assign_vars(args, symres; invcheck)

Get the expression of assigning symres to args.

NiLangCore.bcast_assign_vars — Method.

The broadcast version of assign_vars

NiLangCore.compile_ex — Method.

translate to normal julia code.

NiLangCore.dual_ex — Method.

dual_ex(ex)

Get the dual expression of ex.

NiLangCore.struct2namedtuple — Method.

convert an object to a named tuple.

Instructions

NiLang.CONJ — Method.

CONJ(a!) -> a!'

source

NiLang.DIVINT — Method.

DIVINT(a!, b::Integer) -> a!/b, b

source

NiLang.IROT — Method.

IROT(a!, b!, θ) -> ROT(a!, b!, -θ)

source

NiLang.MULINT — Method.

MULINT(a!, b::Integer) -> a!*b, b

source

NiLang.NEG — Method.

NEG(a!) -> -a!

source

NiLang.ROT — Method.

ROT(a!, b!, θ) -> a!', b!', θ

\[\begin{align} {\rm ROT}(a!, b!, \theta) = \begin{bmatrix} \cos(\theta) & - \sin(\theta)\\ \sin(\theta) & \cos(\theta) \end{bmatrix} \begin{bmatrix} a!\\ b! \end{bmatrix}, \end{align}\]

source

NiLang.SWAP — Method.

SWAP(a!, b!) -> b!, a!

source

NiLang.XOR — Method.

XOR(a!, b) -> a! ⊻ b, b

source

NiLang.arshift — Method.

arshift(x, n)

right shift, sign extending.

source

NiLang.plshift — Method.

plshift(x, n)

periodic left shift.

source

NiLang.prshift — Method.

plshift(x, n)

periodic right shift.

source

NiLang.rot — Method.

rot(a, b, θ)

rotate variables a and b by an angle θ

source

NiLang.AutoBcast — Type.

AutoBcast{T} <: IWrapper{T}

A vectorized variable.

source

Automatic Differentiation

NiLang.AD.@nograd — Macro.

@nograd f(args...)

Mark f(args...) as having no gradients.

source

NiLang.AD.beijingring! — Method.

beijingring!(x::AbstractFloat)

Allocated a new Beijing ring, it will allocate memory on a global tape NiLang.AD.rings.

source

NiLang.AD.check_grad — Method.

check_grad(f, args; kwargs=(), atol::Real=1e-8, verbose::Bool=false)

Return true if the gradient of f(args..., kwargs...) is reversible.

source

NiLang.AD.collect_hessian — Method.

collect_hessian()

Collect hessian data to a matrix.

source

NiLang.AD.grad — Method.

grad(var)

Get the gradient field of var.

source

NiLang.AD.gradient_numeric — Method.

gradient_numeric(f, args; kwargs=())

Numeric differentiating f(args..., kwargs...).

source

NiLang.AD.hdata — Method.

hdata(h::Tuple{BeijingRing,BeijingRing})

Get he hessian element of w.r.t. two variables.

source

NiLang.AD.hessian_repeat — Method.

hessian_repeat(f, args::Tuple; kwargs=())

Obtain the Hessian matrix of f(args..., kwargs...) by differentiating the first order gradients.

source

NiLang.AD.jacobian — Method.

jacobian(f, args; kwargs=())

Get the Jacobian matrix for function f(args..., kwargs...) by use vectorized variables in the gradient field.

source

NiLang.AD.jacobian_repeat — Method.

jacobian_repeat(f, args; kwargs=())

Get the Jacobian matrix for function f(args..., kwargs...) by computing one row (gradients) a time.

source

NiLang.AD.nrings — Method.

nrings(T)

Number of rings of type T in global storage.

source

NiLang.AD.rings_init! — Method.

rings_init!(T)

Remove all rings of type T in global storage.

source

NiLang.AD.BeijingRing — Type.

BeijingRing{T}

The type to access global Hessian data.

source

NiLang.AD.GVar — Type.

GVar{T,GT} <: IWrapper{T}
GVar(x)

Attach a gradient field to x.

source

NiLang.AD.Loss — Type.

Loss{T} <: IWrapper{T}
Loss(x)

Wrapper used to mark the loss variable.

source

NiLang.AD.NoGrad — Type.

NoGrad{T} <: IWrapper{T}
NoGrad(x)

A NoGrad(x) is equivalent to GVar^{-1}(x), which cancels the GVar wrapper.

source